requirements.txt
python-dotenv
streamlit
youtube_transcript_api
groq
openai
google-generativeai